CLN: create core/sorting.py
