+++ pandas/core/frame.py
@@ -3141,7 +3141,7 @@ class DataFrame(NDFrame):
         -------
         duplicated : Series
         """
-        from pandas.core.groupby import get_group_index
+        from pandas.core.sorting import get_group_index
         from pandas.hashtable import duplicated_int64, _SIZE_HINT_LIMIT
 
         def f(vals):
@@ -3179,7 +3179,7 @@ class DataFrame(NDFrame):
             raise ValueError('Length of ascending (%d) != length of by (%d)' %
                              (len(ascending), len(by)))
         if len(by) > 1:
-            from pandas.core.groupby import _lexsort_indexer
+            from pandas.core.sorting import _lexsort_indexer
 
             def trans(v):
                 if needs_i8_conversion(v):
@@ -3197,7 +3197,7 @@ class DataFrame(NDFrame):
                                        na_position=na_position)
             indexer = _ensure_platform_int(indexer)
         else:
-            from pandas.core.groupby import _nargsort
+            from pandas.core.sorting import _nargsort
 
             by = by[0]
             k = self.xs(by, axis=other_axis).values
@@ -3300,7 +3300,7 @@ class DataFrame(NDFrame):
                                                  sort_remaining=sort_remaining)
 
         elif isinstance(labels, MultiIndex):
-            from pandas.core.groupby import _lexsort_indexer
+            from pandas.core.sorting import _lexsort_indexer
 
             # make sure that the axis is lexsorted to start
             # if not we need to reconstruct to get the correct indexer
@@ -3310,7 +3310,7 @@ class DataFrame(NDFrame):
             indexer = _lexsort_indexer(labels.labels, orders=ascending,
                                        na_position=na_position)
         else:
-            from pandas.core.groupby import _nargsort
+            from pandas.core.sorting import _nargsort
 
             # GH11080 - Check monotonic-ness before sort an index
             # if monotonic (already sorted), return None or copy() according
