+++ pandas/core/series.py
@@ -1786,10 +1786,10 @@ class Series(base.IndexOpsMixin, strings.StringAccessorMixin,
             new_index, indexer = index.sortlevel(level, ascending=ascending,
                                                  sort_remaining=sort_remaining)
         elif isinstance(index, MultiIndex):
-            from pandas.core.groupby import _lexsort_indexer
+            from pandas.core.sorting import _lexsort_indexer
             indexer = _lexsort_indexer(index.labels, orders=ascending)
         else:
-            from pandas.core.groupby import _nargsort
+            from pandas.core.sorting import _nargsort
             indexer = _nargsort(index, kind=kind, ascending=ascending,
                                 na_position=na_position)
